Move GraphQL resolver into dedicated file
